Move target-specific code (i.e., code implementing target handlers and top-level handler helper functions) to a new targets/ dir. We now have targets/linux/deb for debian/ubuntu, targets/linux/rpm for azlinux, and targets/windows for windows.
Move packaging specific code -- for example, code for building up deb and rpm spec files + utility functions for creating buildroot state to a new packaging/ dir.
The only code that remains under frontend/ is specific to the frontend itself, i.e. the frontend entrypoint, target routing, helpers that are specific to all
targets, helpers for solve requests to buildkit, etc.
