+++ cmd/frontend/main.go
@@ -6,10 +6,10 @@ import (
 
 	"github.com/Azure/dalec/frontend"
 	"github.com/Azure/dalec/frontend/debug"
-	"github.com/Azure/dalec/frontend/linux/deb/debian"
-	"github.com/Azure/dalec/frontend/linux/deb/ubuntu"
-	"github.com/Azure/dalec/frontend/linux/rpm/azlinux"
-	"github.com/Azure/dalec/frontend/windows"
+	"github.com/Azure/dalec/targets/linux/deb/debian"
+	"github.com/Azure/dalec/targets/linux/deb/ubuntu"
+	"github.com/Azure/dalec/targets/linux/rpm/azlinux"
+	"github.com/Azure/dalec/targets/windows"
 	"github.com/moby/buildkit/frontend/gateway/grpcclient"
 	"github.com/moby/buildkit/util/appcontext"
 	"github.com/moby/buildkit/util/bklog"
