+++ packaging/linux/deb/templates/debian_rules.tmpl
