+++ packaging/linux/rpm/sources.go
