+++ packaging/linux/rpm/template.go
