+++ packaging/linux/rpm/template_test.go
