+++ targets/linux/deb/debian/bookworm.go
@@ -1,7 +1,7 @@
 package debian
 
 import (
-	"github.com/Azure/dalec/frontend/linux/deb/distro"
+	"github.com/Azure/dalec/targets/linux/deb/distro"
 )
 
 const (
