+++ targets/linux/deb/ubuntu/bionic.go
@@ -1,7 +1,7 @@
 package ubuntu
 
 import (
-	"github.com/Azure/dalec/frontend/linux/deb/distro"
+	"github.com/Azure/dalec/targets/linux/deb/distro"
 )
 
 const (
