+++ targets/linux/rpm/azlinux/common.go
