+++ targets/linux/rpm/azlinux/mariner2.go
@@ -1,7 +1,7 @@
 package azlinux
 
 import (
-	"github.com/Azure/dalec/frontend/linux/rpm/distro"
+	"github.com/Azure/dalec/targets/linux/rpm/distro"
 )
 
 const (
