+++ targets/linux/rpm/distro/debug.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/Azure/dalec"
 	"github.com/Azure/dalec/frontend"
-	"github.com/Azure/dalec/frontend/linux/rpm"
+	"github.com/Azure/dalec/packaging/linux/rpm"
 	"github.com/containerd/platforms"
 	"github.com/moby/buildkit/client/llb"
 	gwclient "github.com/moby/buildkit/frontend/gateway/client"
