+++ targets/linux/rpm/distro/worker.go
