+++ targets/windows/handle_container.go
