+++ targets/windows/validation.go
