+++ test/windows_test.go
@@ -7,8 +7,8 @@ import (
 	"testing"
 
 	"github.com/Azure/dalec"
-	"github.com/Azure/dalec/frontend/linux/deb/ubuntu"
-	"github.com/Azure/dalec/frontend/windows"
+	"github.com/Azure/dalec/targets/linux/deb/ubuntu"
+	"github.com/Azure/dalec/targets/windows"
 	"github.com/moby/buildkit/client/llb"
 	gwclient "github.com/moby/buildkit/frontend/gateway/client"
 	moby_buildkit_v1_frontend "github.com/moby/buildkit/frontend/gateway/pb"
